+++ package.json
@@ -4,8 +4,8 @@
   "description": "{ Project }",
   "main": "index.js",
   "scripts": {
-    "build": "node_modules/.bin/nodemon --exec babel-node --debug webpack/production.js --watch webpack/production.js",
-    "dev": "node_modules/.bin/babel-node webpack/development.js",
+    "build": "node_modules/.bin/nodemon --exec babel-node --debug toolkit/webpack/production.js --watch webpack/production.js",
+    "dev": "node_modules/.bin/babel-node toolkit/webpack/development.js",
     "lint": "eslint src"
   },
   "repository": {
