+++ toolkit/webpack/production.js
@@ -2,7 +2,7 @@ import webpack from 'webpack';
 import config from './config';
 
 // Use the express production server
-import '../src/server/app';
+import '../../src/server/app';
 
 // compile all files necessary for serving
 const compiler = webpack(config);
